Move brig's internal servant api to wire-api.
(will need this in a moment to generate servant-clients in galley
integration tests.)
