+++ libs/brig-types/brig-types.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 2b3c91a37e09e61ca32cf32ab80c46e8f019926c78165bec82d9c84a5b8d20a6
+-- hash: 42d648b07cd5a5d45e2ca36fd4631c37047fc74aea88fce77073a3efa99327a2
 
 name:           brig-types
 version:        1.35.0
@@ -40,7 +40,6 @@ library
       Brig.Types.Test.Arbitrary
       Brig.Types.User
       Brig.Types.User.Auth
-      Brig.Types.User.EJPD
       Brig.Types.User.Event
   other-modules:
       Paths_brig_types
