+++ libs/wire-api/package.yaml
@@ -13,6 +13,7 @@ dependencies:
 - containers >=0.5
 - imports
 - types-common >=0.16
+- servant-swagger-ui
 library:
   source-dirs: src
   dependencies:
