+++ services/brig/src/Brig/API/Internal.hs
@@ -19,8 +19,8 @@ module Brig.API.Internal
   ( sitemap,
     servantSitemap,
     swaggerDocsAPI,
-    ServantAPI,
-    SwaggerDocsAPI,
+    BrigIRoutes.API,
+    BrigIRoutes.SwaggerDocsAPI,
   )
 where
 
@@ -42,13 +42,12 @@ import Brig.Team.DB (lookupInvitationByEmail)
 import Brig.Types
 import Brig.Types.Intra
 import Brig.Types.Team.LegalHold (LegalHoldClientRequest (..))
-import qualified Brig.Types.User.EJPD as EJPD
 import Brig.Types.User.Event (UserEvent (UserUpdated), UserUpdatedData (eupSSOId, eupSSOIdRemoved), emptyUserUpdatedData)
 import qualified Brig.User.API.Auth as Auth
 import qualified Brig.User.API.Search as Search
 import qualified Brig.User.EJPD
 import Control.Error hiding (bool)
-import Control.Lens (view, (.~))
+import Control.Lens (view)
 import Data.Aeson hiding (json)
 import Data.ByteString.Conversion
 import qualified Data.ByteString.Conversion as List
@@ -57,7 +56,6 @@ import Data.Id as Id
 import qualified Data.List1 as List1
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
-import Data.Swagger (HasInfo (info), HasTitle (title), Swagger)
 import Galley.Types (UserClients (..))
 import Imports hiding (head)
 import Network.HTTP.Types.Status
@@ -67,12 +65,11 @@ import Network.Wai.Routing
 import Network.Wai.Utilities as Utilities
 import Network.Wai.Utilities.ZAuth (zauthConnId, zauthUserId)
 import Servant hiding (Handler, JSON, addHeader, respond)
-import qualified Servant
-import Servant.Swagger (HasSwagger (toSwagger))
 import Servant.Swagger.Internal.Orphans ()
 import Servant.Swagger.UI
 import qualified System.Logger.Class as Log
 import Wire.API.ErrorDescription
+import qualified Wire.API.Routes.Internal.Brig as BrigIRoutes
 import qualified Wire.API.Team.Feature as ApiFt
 import Wire.API.User
 import Wire.API.User.Client (UserClientsFull (..))
@@ -81,61 +78,7 @@ import Wire.API.User.RichInfo
 ---------------------------------------------------------------------------
 -- Sitemap (servant)
 
-type EJPDRequest =
-  Summary
-    "Identify users for law enforcement.  Wire has legal requirements to cooperate \
-    \with the authorities.  The wire backend operations team uses this to answer \
-    \identification requests manually.  It is our best-effort representation of the \
-    \minimum required information we need to hand over about targets and (in some \
-    \cases) their communication peers.  For more information, consult ejpd.admin.ch."
-    :> "ejpd-request"
-    :> QueryParam'
-         [ Optional,
-           Strict,
-           Description "Also provide information about all contacts of the identified users"
-         ]
-         "include_contacts"
-         Bool
-    :> Servant.ReqBody '[Servant.JSON] EJPD.EJPDRequestBody
-    :> Post '[Servant.JSON] EJPD.EJPDResponseBody
-
-type GetAccountFeatureConfig =
-  Summary
-    "Read access to cassandra field `brig.user.feature_conference_calling"
-    :> "users"
-    :> Capture "uid" UserId
-    :> "features"
-    :> "conferenceCalling"
-    :> Get '[Servant.JSON] (ApiFt.TeamFeatureStatus 'ApiFt.TeamFeatureConferenceCalling)
-
-type PutAccountFeatureConfig =
-  Summary
-    "Read access to cassandra field `brig.user.feature_conference_calling"
-    :> "users"
-    :> Capture "uid" UserId
-    :> "features"
-    :> "conferenceCalling"
-    :> Servant.ReqBody '[Servant.JSON] (ApiFt.TeamFeatureStatus 'ApiFt.TeamFeatureConferenceCalling)
-    :> Get '[Servant.JSON] NoContent
-
-type DeleteAccountFeatureConfig =
-  Summary
-    "Read access to cassandra field `brig.user.feature_conference_calling"
-    :> "users"
-    :> Capture "uid" UserId
-    :> "features"
-    :> "conferenceCalling"
-    :> Delete '[Servant.JSON] NoContent
-
-type ServantAPI =
-  "i"
-    :> ( EJPDRequest
-           :<|> GetAccountFeatureConfig
-           :<|> PutAccountFeatureConfig
-           :<|> DeleteAccountFeatureConfig
-       )
-
-servantSitemap :: ServerT ServantAPI Handler
+servantSitemap :: ServerT BrigIRoutes.API Handler
 servantSitemap =
   Brig.User.EJPD.ejpdRequest
     :<|> getAccountFeatureConfig
@@ -155,15 +98,8 @@ deleteAccountFeatureConfig :: UserId -> Handler NoContent
 deleteAccountFeatureConfig uid =
   lift $ Data.updateFeatureConferenceCalling uid Nothing $> NoContent
 
-type SwaggerDocsAPI = "api" :> "internal" :> SwaggerSchemaUI "swagger-ui" "swagger.json"
-
-swaggerDocsAPI :: Servant.Server SwaggerDocsAPI
-swaggerDocsAPI = swaggerSchemaUIServer swaggerDoc
-
-swaggerDoc :: Swagger
-swaggerDoc =
-  toSwagger (Proxy @ServantAPI)
-    & info . title .~ "Wire-Server API as Swagger 2.0 (internal end-points; incomplete) "
+swaggerDocsAPI :: Servant.Server BrigIRoutes.SwaggerDocsAPI
+swaggerDocsAPI = swaggerSchemaUIServer BrigIRoutes.swaggerDoc
 
 ---------------------------------------------------------------------------
 -- Sitemap (wai-route)
